Mejoras en gestión de Magnets
